+++ ewm-service/src/main/java/ru/practicum/ewmservice/event/dto/EventRequestStatusUpdateRequest.java
@@ -10,7 +10,8 @@ import java.util.List;
 @Setter
 public class EventRequestStatusUpdateRequest {
 
-    private List<Integer> requestIds;
+
+    private List<Long> requestIds;
 
     private Status status;
 }
